qa/workunits/rbd/copy.sh: skip some tests when tiering is enabled
The rados ls bit doesn't work.

+++ qa/workunits/rbd/copy.sh
@@ -5,6 +5,11 @@ rbd ls | wc -l | grep -v '^0$' && echo "nonempty rbd pool, aborting!  run this s
 
 IMGS="testimg1 testimg2 testimg3 foo foo2 bar bar2 test1 test2 test3"
 
+tiered=0
+if ceph osd dump | grep ^pool | grep "'rbd'" | grep tier; then
+    tiered=1
+fi
+
 remove_images() {
     for img in $IMGS
     do
@@ -177,20 +182,22 @@ test_remove() {
     rbd rm test1
     rbd ls | wc -l | grep "^0$"
 
-    # remove with header missing
-    rbd create --new-format -s 1 test2
-    HEADER=$(rados -p rbd ls | grep '^rbd_header')
-    rados -p rbd rm $HEADER
-    rbd rm test2
-    rbd ls | wc -l | grep "^0$"
+    if [ $tiered -eq 0 ]; then
+        # remove with header missing
+	rbd create --new-format -s 1 test2
+	HEADER=$(rados -p rbd ls | grep '^rbd_header')
+	rados -p rbd rm $HEADER
+	rbd rm test2
+	rbd ls | wc -l | grep "^0$"
 
-    # remove with header and id missing
-    rbd create --new-format -s 1 test2
-    HEADER=$(rados -p rbd ls | grep '^rbd_header')
-    rados -p rbd rm $HEADER
-    rados -p rbd rm rbd_id.test2
-    rbd rm test2
-    rbd ls | wc -l | grep "^0$"
+        # remove with header and id missing
+	rbd create --new-format -s 1 test2
+	HEADER=$(rados -p rbd ls | grep '^rbd_header')
+	rados -p rbd rm $HEADER
+	rados -p rbd rm rbd_id.test2
+	rbd rm test2
+	rbd ls | wc -l | grep "^0$"
+    fi
 
     # remove with rbd_children object missing (and, by extension,
     # with child not mentioned in rbd_children)
